".gitignore" file: update file.
Currently, only the frontend "node_modules" directory is being
ignored.

This means that the "node_modules" directory for the user-service and
matching-service is still being tracked as a commit to the main
repo.

Let's ensure that the "node_modules" directory for the user-service
and matching-service are also ignored.

The necessary lines to ignore both directories have been added to
the ".gitignore" file.