"MatchingPage.js" file: created file in "components" directory. "App.js" file: updated file. "Navbar.js" file: created file in "components" directory. "index.css" file: updated file.
Currently, there is not matching page that exists that provides the
user with the ability to input their desired difficulty.

We require this information to give users questions of the
appropriate difficulty.

Let's create a page that allows users to select their desired
difficulty level of questions.

This page is to be supported by our frontend microservice and the
creation of this page will be done using React.

The route to this service has also been created in "App.js". To
view the page on your browser, enter "npm run start" in the terminal
and access "http://localhost:3000/matching".

The page currently allows users to select a difficulty and confirm
their decision.

The page also contains a navigation bar that allows users to logout.
It currently returns the user back to the sign up page, a behavior
that will be changed in future updates.

The "index.css" file has also been updated to properly format the
Matching page. This provides visual clarity to the users.