"MatchingPage.js" file: updated file. "App.js" file: updated file. "MatchingRoom.js" file: created file in "components" directory. "index.js" file: updated file found in "matching-service" directory.
Currently, the frontend of the matching-service does not emit match
event, handle matchSuccess event or handle matchFail event.

We need these to be done so that the matching-service frontend can
actually communicate with the backend to fulfill the purpose of the
matching-service.

Let's handle these individually.

Firstly, the "MatchingPage.js" has been updated to create a socket
on port 3000 and connect to the socket of the backend server that
is on port 8001.

Secondly, the "MatchingPage.js" has been updated to emit a 'match'
event whenever the user clicks the 'confirmation' button.

Thirdly, the "MatchingPage.js" has been updated to listen on port
8001 for 'matchSuccess' event. Then the frontend navigates to the
matching room with the corresponding matchedRoomId.
The "MatchingRoom.js" file was created to facilitate this.

Fourthly, the "MatchingPage.js" has been updated to listen on port
8001 for 'matchFail' event. Then the 'reset' and 'repeat' buttons
are revealed to indicate to the user that the matching has failed.
These buttons allow the user to change to another difficulty and
try to rematch respectively.

Additional point 1: the console log when a matchSuccess event is
emitted has been updated in the "index.js" file in the
"matching-service" directory to reveal more information for testing
debugging purposes.

Additional point 2: currently, the usernames used in the
matching-service is a randomly generated string of numbers using
Math.random().toString(). This will be updated in the future once we
properly navigate from the HomePage to the MatchingPage.