Update MatchingRoom.js
Changes are specific to implementing communication service.

The chatbox now has the basic function of taking a users message and
logging it to the console when the user clicks "send". The input
field is automatically cleared for further inputs.

Changes made:
Add "input" state which keeps track of the most recent input.
Add handleChange() which is called whenever the user types anything
into the "input" field.
Add handleSend() which gets called when the user clicks "send"; it
logs the "input" into the console and clears the "input" state.