Update index.js in matching-service directory. Add matching-service-config.js and matching-service-constants.js directory.
Currently, when the matching service backend server emits a match
success event it only transfers the roomId that the users are going
into.

Let's include the question that the 2 users will be doing.

To do this the matching service backend server will need to make a
get request to the question service backend. This is done using the
axios framework.

In addition, the config and constants .js files were added to as
additional files to support the new functionalities.