Update MatchingPage.js in the frontend directory.
Due to the asynchronous behavior of the React framework, there is a
need to implement a work around to make use of states immediately
after a change to those states.

We will make use of the useEffect() functions that is a part of the
React framework.

The initial render of the matching page is also being tracked now.
This so we have more control over what useEffects() are ran before
and after the initial render.

The matching page now accepts a question in the location.state
from the matching service. This will be used when navigating from
the matching page to the matching room.